# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AqarGo API",
        "description": "AqarGo real estate platform API documentation",
        "contact": {
            "email": "support@aqargo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/profile/getMyProfile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get the authenticated user's profile",
                "operationId": "db28461030543940c6a698455e9ab6be",
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "401": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/show/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Show another user's profile",
                "operationId": "1604e81f23f9a108f1910010507e465a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "404": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/create": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Create a user profile",
                "operationId": "3a6e92ec358def9e9e7fbc3456617226",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ahmad"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Ali"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0912345678"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 33.51
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 36.27
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Damascus, Syria"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update the authenticated user profile",
                "operationId": "658523cb93a6af7d87e1ca8740b514f5",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ahmad"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Ali"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0912345678"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 33.51
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 36.27
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Damascus, Syria"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/delete": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete the authenticated user profile",
                "operationId": "785571e5bd10f54bc9556293d98cf3e2",
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation Error."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your JWT token"
            }
        }
    },
...
}